requirements.txt
python-dotenv
aiogram==2.5

streamlit
streamlit_chat

logging
warnings

pandas
numpy

google-generativeai
langchain
langchain_community
langchain_core
langchain_google_genai

sentence-transformers==2.2.2